chore(harness): complete unified agent cutover
Refs: SAP-3152
