--- .changeset/plan-agents-label.md
@@ -1,6 +1,0 @@
----
-"@sapiom/harness": patch
-"@sapiom/harness-desktop": patch
----
-
-Rename the pinned Agent Map tab in the Studio project sidebar to Plan Agents so its purpose is clear. The underlying Agent Map view and behavior are unchanged.
