--- .changeset/planner-owned-agent-creation.md
@@ -1,6 +1,0 @@
----
-"@sapiom/harness": minor
-"@sapiom/harness-desktop": patch
----
-
-Make Agent Map planning the only agent-creation route in Studio projects. Empty sidebar rows no longer offer to create the first agent directly, and project menus no longer offer direct create or in-session scaffold actions.
