+++ docs/plans/agent-studio-plan-first-agent-map/journey-roadmap.json
@@ -1,0 +1,19 @@
+{
+  "schemaVersion": 1,
+  "authority": "SAP-3147",
+  "checkpoint": "SAP-3152",
+  "currentStack": [
+    { "issue": "SAP-3148", "pr": 804, "head": "5d00c55925d1907670bed6e885384b25eb775a73", "outcome": "unified identity, bootstrap, and navigation" },
+    { "issue": "SAP-3149", "pr": 806, "head": "90eb569eb6b90b917c49128b6a23c7215a0843d6", "outcome": "neutral immutable versions and universal plan authoring" },
+    { "issue": "SAP-3150", "pr": 807, "head": "acb2dbae6a3e533e01a065c43fa4109bdd82ca14", "outcome": "role-neutral focused briefs" },
+    { "issue": "SAP-3151", "pr": 808, "head": "ca747232e0b775b5be5c69c5427c778b3774dd4f", "outcome": "writable idempotent nested delegation", "review": "user-directed exception recorded in cutover ledger" },
+    { "issue": "SAP-3152", "pr": null, "head": null, "outcome": "verification, documentation, and reconciliation" }
+  ],
+  "dependencyOrder": ["SAP-3148", "SAP-3149", "SAP-3150", "SAP-3151", "SAP-3152"],
+  "historicalFoundations": ["E0", "E1", "E2"],
+  "laterWork": [
+    { "epic": "E6", "directive": "consume shared neutral map, plan, brief, and session context" },
+    { "epic": "E7", "directive": "reconcile evidence without silently changing intent" },
+    { "epic": "E8", "directive": "adopt existing projects and remove obsolete authority without changing the per-agent graph" }
+  ]
+}
