+++ packages/harness/src/core/agent-map-aggregate-migration.ts
@@ -9,10 +9,10 @@ import {
   parseAgentMapProposalReceipt,
   parseMapChangeProposal,
   parseMapOperation,
-  parseLegacyE2ProposalActor,
   parseProjectAgentActorRef,
   type PersistedAgentMapProposalReceipt,
 } from "../shared/agent-map-codec.js";
+import { parseLegacyE2ProposalActor } from "../shared/agent-map-legacy-migration.js";
 import { parseAgentMapVersion } from "../shared/agent-map-version-codec.js";
 import { canonicalDigest, canonicalJson } from "../shared/agent-map-canonical.js";
 import type {
