+++ packages/harness/src/server/agent-map.test.ts
@@ -8,27 +8,10 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import { AgentMapWorkspaceStore } from "../core/agent-map-workspace-store.js";
 import { StudioProjectCatalog } from "../core/studio-project-catalog.js";
 import { StudioWorkspacePreferenceStore } from "../core/studio-workspace-preferences.js";
-import {
-  ProjectBootstrapInputCapacityError,
-  ProjectBootstrapRequestIdConflictError,
-  ProjectBootstrapRetryUnavailableError,
-  type ProjectBootstrapCoordinator,
-} from "../core/planner-greeting.js";
-import {
-  ProjectSessionScopeUnavailableError,
-  SessionBackgroundInputPreemptedError,
-  SessionInputIsolationError,
-  SessionNotReadyError,
-} from "../core/session-manager.js";
-import {
-  ProjectSessionError,
-  type ProjectSessionService,
-} from "../core/planning-session.js";
 import type {
   AgentMapWorkspaceResponse,
   StudioProjectSummary,
 } from "../shared/agent-map.js";
-import type { HarnessSession } from "../shared/types.js";
 import { createBootTokenMiddleware } from "./auth.js";
 import { createAgentMapRouter } from "./agent-map.js";
 
@@ -48,25 +31,6 @@ describe("createAgentMapRouter", () => {
   });
 
   async function start(projectLifecycle?: {
-    projectSessions?: ProjectSessionService;
-    projectBootstrap?: ProjectBootstrapCoordinator;
-    submitSessionInput?: (
-      sessionId: string,
-      text: string,
-      submit: boolean,
-      requestId?: string,
-    ) => Promise<
-      | boolean
-      | {
-          ok: boolean;
-          receipt?: {
-            requestId: string | null;
-            inputId: string;
-            status: "queued" | "submitted" | "uncertain" | "completed";
-            acceptedAt: string;
-          };
-        }
-    >;
     onProjectCreated?: (project: StudioProjectSummary) => Promise<void> | void;
     onRootBound?: (
       project: StudioProjectSummary,
@@ -117,9 +81,6 @@ describe("createAgentMapRouter", () => {
         isWorkflowScanComplete: () => true,
         listWorkspaceScopes,
         ...projectLifecycle,
-        ...(projectLifecycle && !projectLifecycle.submitSessionInput
-          ? { submitSessionInput: async () => true }
-          : {}),
       }),
     );
     server = app.listen(0);
@@ -589,553 +550,4 @@ describe("createAgentMapRouter", () => {
     expect(await fs.readFile(workspacePath, "utf8")).toBe("{bad-json");
   });
 
-  it("keeps legacy planner routes as project-scoped aliases to neutral services", async () => {
-    let fixtureProjectId = "";
-    const plannerSession = {
-      id: "planner-session-1",
-      agentSessionId: null,
-      harness: "codex",
-      cwd: "/server/private/project",
-      title: "project",
-      status: "running",
-      createdAt: "2026-09-01T00:00:00.000Z",
-      lastActiveAt: "2026-09-01T00:00:00.000Z",
-      exitCode: null,
-      boundWorkflowPath: null,
-      ready: false,
-    } as HarnessSession;
-    const open = vi.fn(async () => ({
-      session: plannerSession,
-      resolution: "created" as const,
-    }));
-    const requireOwned = vi.fn(() => plannerSession);
-    const enqueue = vi.fn(async (_sessionId: string, _text: string) => {
-      const metadata = {
-        projectId: fixtureProjectId,
-        userId: "user-1",
-        targetSessionId: plannerSession.id,
-        bootstrap: {
-          status: "skipped" as const,
-          reason: "user-proceeded" as const,
-        },
-        queuedInputIds: ["input-1"],
-      };
-      plannerSession.projectBootstrap = metadata;
-      return metadata;
-    });
-    const retry = vi.fn(async () => {
-      if (!plannerSession.projectBootstrap) {
-        throw new Error("missing project bootstrap metadata");
-      }
-      plannerSession.projectBootstrap = {
-        ...plannerSession.projectBootstrap,
-        bootstrap: { status: "generating", attemptId: "attempt-2" },
-      };
-    });
-    // This is the one canonical ordinary-session input authority shared with
-    // POST /sessions/:id/input. The compatibility route must call it even
-    // while bootstrap owns the durable FIFO; it may not enqueue independently.
-    const submitSessionInput = vi.fn(
-      async (
-        sessionId: string,
-        text: string,
-        _submit: boolean,
-        requestId?: string,
-      ) => {
-        await enqueue(sessionId, text);
-        return requestId
-          ? {
-              ok: true as const,
-              receipt: {
-                requestId,
-                inputId: "input-1",
-                status: "queued" as const,
-                acceptedAt: "2026-09-04T00:00:00.000Z",
-              },
-            }
-          : true;
-      },
-    );
-    const fixture = await start({
-      projectSessions: {
-        open,
-        requireOwned,
-      } as unknown as ProjectSessionService,
-      projectBootstrap: {
-        enqueue,
-        retry,
-      } as unknown as ProjectBootstrapCoordinator,
-      submitSessionInput,
-    });
-    fixtureProjectId = fixture.project.projectId;
-    plannerSession.agentMapIdentity = {
-      projectId: fixtureProjectId,
-      sessionId: plannerSession.id,
-      userId: "user-1",
-    };
-    plannerSession.projectBootstrap = {
-      projectId: fixtureProjectId,
-      targetSessionId: plannerSession.id,
-      userId: "user-1",
-      bootstrap: {
-        status: "failed",
-        retryable: true,
-        errorCode: "model_turn_failed",
-      },
-      queuedInputIds: [],
-    };
-    const route = `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions`;
-
-    expect(
-      (
-        await fetch(route, {
-          method: "POST",
-          headers: { "content-type": "application/json" },
-          body: JSON.stringify({ mode: "fresh" }),
-        })
-      ).status,
-    ).toBe(401);
-    const forged = await fetch(route, {
-      method: "POST",
-      headers: {
-        "content-type": "application/json",
-        "X-Harness-Token": "test-token",
-      },
-      body: JSON.stringify({
-        mode: "fresh",
-        role: "map-planner",
-        projectId: fixture.project.projectId,
-      }),
-    });
-    expect(forged.status).toBe(400);
-    expect(open).not.toHaveBeenCalled();
-
-    const valid = await fetch(route, {
-      method: "POST",
-      headers: {
-        "content-type": "application/json",
-        "X-Harness-Token": "test-token",
-      },
-      body: JSON.stringify({ mode: "fresh", harness: "codex" }),
-    });
-    expect(valid.status).toBe(201);
-    expect(open).toHaveBeenCalledWith(fixture.project.projectId, {
-      mode: "fresh",
-      harness: "codex",
-    });
-
-    const message = await fetch(`${route}/${plannerSession.id}/messages`, {
-      method: "POST",
-      headers: {
-        "content-type": "application/json",
-        "X-Harness-Token": "test-token",
-      },
-      body: JSON.stringify({
-        text: "Build a support triage system",
-        requestId: "request-build-support",
-      }),
-    });
-    expect(message.status).toBe(202);
-    expect(await message.json()).toEqual({
-      metadata: {
-        projectId: fixture.project.projectId,
-        targetSessionId: plannerSession.id,
-        userId: "user-1",
-        bootstrap: { status: "skipped", reason: "user-proceeded" },
-        queuedInputIds: ["input-1"],
-      },
-      receipt: {
-        requestId: "request-build-support",
-        inputId: "input-1",
-        status: "queued",
-        acceptedAt: "2026-09-04T00:00:00.000Z",
-      },
-    });
-    expect(requireOwned).toHaveBeenCalledWith(
-      fixture.project.projectId,
-      plannerSession.id,
-    );
-    expect(enqueue).toHaveBeenCalledWith(
-      plannerSession.id,
-      "Build a support triage system",
-    );
-    expect(submitSessionInput).toHaveBeenCalledWith(
-      plannerSession.id,
-      "Build a support triage system",
-      true,
-      "request-build-support",
-    );
-
-    const followUp = await fetch(`${route}/${plannerSession.id}/messages`, {
-      method: "POST",
-      headers: {
-        "content-type": "application/json",
-        "X-Harness-Token": "test-token",
-      },
-      body: JSON.stringify({ text: "Keep this behind the durable FIFO" }),
-    });
-    expect(followUp.status).toBe(202);
-    expect(enqueue).toHaveBeenLastCalledWith(
-      plannerSession.id,
-      "Keep this behind the durable FIFO",
-    );
-    expect(enqueue).toHaveBeenCalledTimes(2);
-    expect(submitSessionInput).toHaveBeenCalledTimes(2);
-
-    plannerSession.projectBootstrap = {
-      ...plannerSession.projectBootstrap!,
-      bootstrap: {
-        status: "failed",
-        retryable: true,
-        errorCode: "model_turn_failed",
-      },
-      queuedInputIds: [],
-    };
-
-    const retryResponse = await fetch(
-      `${route}/${plannerSession.id}/greeting/retry`,
-      {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-          "X-Harness-Token": "test-token",
-        },
-        body: "{}",
-      },
-    );
-    expect(retryResponse.status).toBe(202);
-    expect(await retryResponse.json()).toEqual({
-      metadata: {
-        ...plannerSession.projectBootstrap,
-        bootstrap: { status: "generating", attemptId: "attempt-2" },
-      },
-    });
-    expect(retry).toHaveBeenCalledWith(plannerSession.id);
-  });
-
-  it("lets an ordinary project session use the legacy message alias without bootstrap metadata", async () => {
-    const ordinarySession = {
-      id: "ordinary-session-1",
-      agentSessionId: "provider-session-1",
-      harness: "codex",
-      cwd: "/server/private/project",
-      title: "Implementation",
-      status: "running",
-      createdAt: "2026-09-01T00:00:00.000Z",
-      lastActiveAt: "2026-09-01T00:00:00.000Z",
-      exitCode: null,
-      boundWorkflowPath: null,
-      ready: true,
-    } as HarnessSession;
-    const requireOwned = vi.fn(async () => ordinarySession);
-    const submitSessionInput = vi.fn(async () => true);
-    const fixture = await start({
-      projectSessions: {
-        open: vi.fn(),
-        requireOwned,
-      } as unknown as ProjectSessionService,
-      submitSessionInput,
-    });
-    ordinarySession.agentMapIdentity = {
-      projectId: fixture.project.projectId,
-      sessionId: ordinarySession.id,
-      userId: "user-1",
-    };
-
-    const response = await fetch(
-      `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/${ordinarySession.id}/messages`,
-      {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-          "X-Harness-Token": "test-token",
-        },
-        body: JSON.stringify({ text: "Continue the implementation" }),
-      },
-    );
-
-    expect(response.status).toBe(202);
-    expect(await response.json()).toEqual({ metadata: null });
-    expect(requireOwned).toHaveBeenCalledWith(
-      fixture.project.projectId,
-      ordinarySession.id,
-    );
-    expect(submitSessionInput).toHaveBeenCalledTimes(1);
-    expect(submitSessionInput).toHaveBeenCalledWith(
-      ordinarySession.id,
-      "Continue the implementation",
-      true,
-      undefined,
-    );
-  });
-
-  it("dispatches each compatibility request once through the canonical authority", async () => {
-    const ordinarySession = {
-      id: "ordinary-session-1",
-      agentSessionId: "provider-session-1",
-      harness: "codex",
-      cwd: "/server/private/project",
-      title: "Implementation",
-      status: "running",
-      createdAt: "2026-09-01T00:00:00.000Z",
-      lastActiveAt: "2026-09-01T00:00:00.000Z",
-      exitCode: null,
-      boundWorkflowPath: null,
-      ready: true,
-    } as HarnessSession;
-    const submitSessionInput = vi.fn(async () => true);
-    const fixture = await start({
-      projectSessions: {
-        open: vi.fn(),
-        requireOwned: vi.fn(async () => ordinarySession),
-      } as unknown as ProjectSessionService,
-      submitSessionInput,
-    });
-    const route = `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/${ordinarySession.id}/messages`;
-    const request = () =>
-      fetch(route, {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-          "X-Harness-Token": "test-token",
-        },
-        body: JSON.stringify({ text: "same visible text" }),
-      });
-
-    const responses = await Promise.all([request(), request()]);
-
-    expect(responses.map((response) => response.status)).toEqual([202, 202]);
-    expect(submitSessionInput).toHaveBeenCalledTimes(2);
-    for (const call of submitSessionInput.mock.calls) {
-      expect(call).toEqual([
-        ordinarySession.id,
-        "same visible text",
-        true,
-        undefined,
-      ]);
-    }
-  });
-
-  it("forwards request IDs and returns the same durable conflict semantics through the compatibility alias", async () => {
-    const ordinarySession = {
-      id: "bootstrap-session-1",
-      status: "running",
-    } as HarnessSession;
-    const submitSessionInput = vi.fn(async () => {
-      throw new ProjectBootstrapRequestIdConflictError();
-    });
-    const fixture = await start({
-      projectSessions: {
-        open: vi.fn(),
-        requireOwned: vi.fn(async () => ordinarySession),
-      } as unknown as ProjectSessionService,
-      submitSessionInput,
-    });
-
-    const response = await fetch(
-      `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/${ordinarySession.id}/messages`,
-      {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-          "X-Harness-Token": "test-token",
-        },
-        body: JSON.stringify({
-          text: "changed payload",
-          requestId: "request-reused",
-        }),
-      },
-    );
-
-    expect(response.status).toBe(409);
-    expect(await response.json()).toEqual({
-      code: "project_bootstrap_request_id_reused",
-      error: "project bootstrap request id was reused with different input",
-    });
-    expect(submitSessionInput).toHaveBeenCalledWith(
-      ordinarySession.id,
-      "changed payload",
-      true,
-      "request-reused",
-    );
-  });
-
-  it("returns bounded durable-input capacity through the compatibility alias", async () => {
-    const ordinarySession = {
-      id: "bootstrap-session-capacity",
-      status: "running",
-    } as HarnessSession;
-    const submitSessionInput = vi.fn(async () => {
-      throw new ProjectBootstrapInputCapacityError();
-    });
-    const fixture = await start({
-      projectSessions: {
-        open: vi.fn(),
-        requireOwned: vi.fn(async () => ordinarySession),
-      } as unknown as ProjectSessionService,
-      submitSessionInput,
-    });
-
-    const response = await fetch(
-      `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/${ordinarySession.id}/messages`,
-      {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-          "X-Harness-Token": "test-token",
-        },
-        body: JSON.stringify({
-          text: "new logical request",
-          requestId: "request-at-capacity",
-        }),
-      },
-    );
-
-    expect(response.status).toBe(409);
-    expect(await response.json()).toEqual({
-      code: "project_bootstrap_input_capacity",
-      error: "project bootstrap input receipt capacity is temporarily full",
-    });
-    expect(submitSessionInput).toHaveBeenCalledWith(
-      ordinarySession.id,
-      "new logical request",
-      true,
-      "request-at-capacity",
-    );
-  });
-
-  it("maps canonical input rejection without retrying compatibility dispatch", async () => {
-    const ordinarySession = {
-      id: "ordinary-session-1",
-      status: "running",
-    } as HarnessSession;
-    const submitSessionInput = vi
-      .fn<
-        (sessionId: string, text: string, submit: boolean) => Promise<boolean>
-      >()
-      .mockResolvedValueOnce(false)
-      .mockRejectedValueOnce(new SessionNotReadyError(ordinarySession.id))
-      .mockRejectedValueOnce(new SessionBackgroundInputPreemptedError(false))
-      .mockRejectedValueOnce(new SessionInputIsolationError());
-    const fixture = await start({
-      projectSessions: {
-        open: vi.fn(),
-        requireOwned: vi.fn(async () => ordinarySession),
-      } as unknown as ProjectSessionService,
-      submitSessionInput,
-    });
-    const route = `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/${ordinarySession.id}/messages`;
-    const request = () =>
-      fetch(route, {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-          "X-Harness-Token": "test-token",
-        },
-        body: JSON.stringify({ text: "one dispatch" }),
-      });
-
-    const missing = await request();
-    const unready = await request();
-    const concurrent = await request();
-    const isolated = await request();
-
-    expect(missing.status).toBe(404);
-    expect(unready.status).toBe(409);
-    expect(await unready.json()).toMatchObject({ code: "SESSION_NOT_READY" });
-    expect(concurrent.status).toBe(409);
-    expect(await concurrent.json()).toEqual({
-      code: "SESSION_BACKGROUND_INPUT_PREEMPTED",
-      error: "background session input was preempted by user input",
-    });
-    expect(isolated.status).toBe(409);
-    expect(await isolated.json()).toEqual({
-      code: "SESSION_INPUT_ISOLATION_REQUIRED",
-      error: "session input is blocked until the terminal composer is reset",
-    });
-    expect(submitSessionInput).toHaveBeenCalledTimes(4);
-  });
-
-  it("rejects foreign planner messages and bounds unavailable retries", async () => {
-    const requireOwned = vi.fn<() => Promise<HarnessSession>>(async () => {
-      throw new ProjectSessionError("forbidden");
-    });
-    const enqueue = vi.fn(async () => ({}) as never);
-    const retry = vi.fn(async () => {
-      throw new ProjectBootstrapRetryUnavailableError();
-    });
-    const fixture = await start({
-      projectSessions: {
-        open: vi.fn(),
-        requireOwned,
-      } as unknown as ProjectSessionService,
-      projectBootstrap: {
-        enqueue,
-        retry,
-      } as unknown as ProjectBootstrapCoordinator,
-    });
-    const headers = {
-      "content-type": "application/json",
-      "X-Harness-Token": "test-token",
-    };
-    const message = await fetch(
-      `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/foreign/messages`,
-      { method: "POST", headers, body: JSON.stringify({ text: "hello" }) },
-    );
-    expect(message.status).toBe(403);
-    expect(await message.json()).toMatchObject({ code: "forbidden" });
-    expect(enqueue).not.toHaveBeenCalled();
-
-    const forbiddenRetry = await fetch(
-      `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/foreign/greeting/retry`,
-      { method: "POST", headers, body: "{}" },
-    );
-    expect(forbiddenRetry.status).toBe(403);
-    expect(await forbiddenRetry.json()).toMatchObject({ code: "forbidden" });
-    expect(retry).not.toHaveBeenCalled();
-
-    requireOwned.mockResolvedValue({ id: "owned" } as HarnessSession);
-    const retryResponse = await fetch(
-      `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/owned/greeting/retry`,
-      { method: "POST", headers, body: "{}" },
-    );
-    expect(retryResponse.status).toBe(409);
-    expect(await retryResponse.json()).toEqual({
-      code: "project_bootstrap_retry_unavailable",
-      error: "project bootstrap retry is not available",
-    });
-  });
-
-  it("returns a bounded compatibility error when project scope cannot be revalidated", async () => {
-    const requireOwned = vi.fn(async () => {
-      throw new ProjectSessionScopeUnavailableError("ordinary-session");
-    });
-    const fixture = await start({
-      projectSessions: {
-        open: vi.fn(),
-        requireOwned,
-      } as unknown as ProjectSessionService,
-      projectBootstrap: {
-        enqueue: vi.fn(),
-        retry: vi.fn(),
-      } as unknown as ProjectBootstrapCoordinator,
-    });
-    const response = await fetch(
-      `${fixture.baseUrl}/api/projects/${fixture.project.projectId}/planner-sessions/ordinary-session/messages`,
-      {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-          "X-Harness-Token": "test-token",
-        },
-        body: JSON.stringify({ text: "continue" }),
-      },
-    );
-
-    expect(response.status).toBe(409);
-    expect(await response.json()).toEqual({
-      code: "PROJECT_SESSION_SCOPE_UNAVAILABLE",
-      error: "the session's Studio project scope could not be revalidated",
-    });
-  });
 });
